+++ example/src/controls/PivotPointMesh.js
