+++ example/src/controls/PointerTracker.js
