+++ example/src/controls/utils.js
@@ -1,0 +1,26 @@
+import { Matrix4 } from 'three';
+
+const _matrix = new Matrix4();
+
+// helper function for constructing a matrix for rotating around a point
+export function makeRotateAroundPoint( point, quat, target ) {
+
+	target.makeTranslation( - point.x, - point.y, - point.z );
+
+	_matrix.makeRotationFromQuaternion( quat );
+	target.premultiply( _matrix );
+
+	_matrix.makeTranslation( point.x, point.y, point.z );
+	target.premultiply( _matrix );
+
+	return target;
+
+}
+
+// get the three.js pointer coords from an event
+export function mouseToCoords( clientX, clientY, element, target ) {
+
+	target.x = ( ( clientX - element.offsetLeft ) / element.clientWidth ) * 2 - 1;
+	target.y = - ( ( clientY - element.offsetTop ) / element.clientHeight ) * 2 + 1;
+
+}
